+++ packages/bridge-controller/src/selectors.ts
@@ -225,9 +225,8 @@ export const selectIsAssetExchangeRateInState = (
 const selectBridgeFeesPerGas = createStructuredBridgeSelector({
   estimatedBaseFeeInDecGwei: ({ gasFeeEstimates }) =>
     gasFeeEstimates?.estimatedBaseFee,
-  maxPriorityFeePerGasInDecGwei: ({ gasFeeEstimates }) =>
-    gasFeeEstimates?.[BRIDGE_PREFERRED_GAS_ESTIMATE]
-      ?.suggestedMaxPriorityFeePerGas,
+  feePerGasInDecGwei: ({ gasFeeEstimates }) =>
+    gasFeeEstimates?.[BRIDGE_PREFERRED_GAS_ESTIMATE]?.suggestedMaxFeePerGas,
   maxFeePerGasInDecGwei: ({ gasFeeEstimates }) =>
     gasFeeEstimates?.high?.suggestedMaxFeePerGas,
 });
@@ -331,8 +330,18 @@ const selectBridgeQuotesWithMetadata = createBridgeSelector(
         toTokenAmount,
         minToTokenAmount,
         swapRate: calcSwapRate(sentAmount.amount, toTokenAmount.amount),
+        /**
+        This is the amount required to submit the transactions
+        Includes the relayer fee or other native fees
+        Should be used for balance checks and tx submission.
+         */
         totalNetworkFee: totalEstimatedNetworkFee,
         totalMaxNetworkFee,
+        /**
+        This contains gas fee estimates for the bridge transaction
+        Does not include the relayer fee (if needed), just the gasLimit and effectiveGas returned by the bridge API
+        Should only be used for display purposes.
+         */
         gasFee,
         adjustedReturn,
         cost,